Load prependers in alphabetical order
Every time Ruby runs `Dir.glob` the sort order is preserved every
time... except the same order is not guaranteed across different
operating systems.

For this reason I think we should sort everything alphabetically when
loading prepeders so that we get the same result across different
operating systems.

The specs check for `ancestors` on the class (what's loaded last is the
first in the array of ancestors).